+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The resulting `.d.ts` are **only** compatible with typescript versions `>= 3.6`.
 ### Added
 
 - Added `currentViewWindow` to get the currently visible area of the graph
+- Added `currentViewWindow` to the 'zoomchange' event
+- Added 'render' event and `onBeforeCompleteRender` callback
 
 ### Fixed